fix cast warning
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@530 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356

+++ driver/linux/omx_main.c
@@ -61,7 +61,7 @@ omx_init(void)
 		goto out_with_net;
 
 	printk(KERN_INFO "Open-MX initialized, using Ethertype 0x%lx\n",
-	       ETH_P_OMX);
+	       (unsigned long) ETH_P_OMX);
 	return 0;
 
  out_with_net:
